+++ bin/assets_cleaner.dart
@@ -1,6 +1,7 @@
 import 'package:args/command_runner.dart';
 import 'package:assets_cleaner/commands/clean_command.dart';
 import 'package:assets_cleaner/commands/init_command.dart';
+import 'package:assets_cleaner/commands/scanlib_command.dart';
 import 'package:assets_cleaner/commands/trash_command.dart';
 import 'package:assets_cleaner/commands/unused_command.dart';
 
@@ -14,5 +15,6 @@ void main(List<String> args) {
     ..addCommand(CleanCommand())
     ..addCommand(InitCommand())
     ..addCommand(TrashCommand())
+    ..addCommand(ScanLibCommand())
     ..run(args);
 }